+++ packages/components/Menu/src/MenuItemCheckbox/useMenuItemCheckbox.ts
@@ -56,6 +56,7 @@ export const useMenuCheckboxInteraction = (
     accessibilityActions,
     accessibilityLabel,
     accessibilityState,
+    accessible,
     componentRef = defaultComponentRef,
     disabled,
     name,
@@ -123,7 +124,7 @@ export const useMenuCheckboxInteraction = (
   return {
     props: {
       ...pressable.props,
-      accessible: true,
+      accessible: accessible ?? true,
       accessibilityActions: accessibilityActionsProp,
       accessibilityLabel,
       accessibilityRole: 'menuitem',